third_party/blink/web_tests/fast/frames/iframe-no-name-expected.txt

Checks that the iframe's name is an empty string if unset.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS frames[0].name is ""
PASS successfullyParsed is true

TEST COMPLETE
